backend/requirements.txt
fastapi==0.115.2
uvicorn==0.22.0
supabase==2.11.0
python-dotenv==1.0.0
python-multipart==0.0.6
appwrite

